refactor: Spark 삭제 및 Spring Controller 분리
