--- src/main/java/chess/application/spark/WebApplication.java
@@ -1,16 +1,0 @@
-package chess.application.spark;
-
-import chess.application.spark.WebChessGame;
-
-import static spark.Spark.port;
-import static spark.Spark.staticFileLocation;
-
-public class WebApplication {
-
-    public static void main(String[] args) {
-        staticFileLocation("/static");
-        port(8080);
-        WebChessGame webChessGame = new WebChessGame();
-        webChessGame.run();
-    }
-}
