+++ src/main/java/chess/application/web/dao/CommandDao.java
@@ -1,4 +1,4 @@
-package chess.application.spring;
+package chess.application.web.dao;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
