--- src/main/resources/templates/start.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>체스 게임: 시작</title>
-</head>
-<body>
-<h1> 체스게임 </h1>
-{{message}}
-<form action="/start" method="post">
-    <input type="submit" value="새로 하기">
-</form>
-<form action="/reload" method="post">
-    <input type="submit" value="이어 하기">
-</form>
-
-</body>
-</html>
